io/config_reader.py: Move YAMLLoader to io.util
